RavenDB-24423 Moved full test suite into StressTests
